@@ -10,7 +10,8 @@
     "lint-fix": "vue-cli-service lint --fix",
     "eslint-fix": "eslint --fix --ext .ts,.vue src",
     "deploy": "yarn lint && yarn build && yarn version --patch && yarn build && firebase deploy",
-    "update-entries": "ts-node src/utils/updateEntries"
+    "update-entries": "ts-node src/utils/updateEntries",
+    "postinstall": "patch-package"
   },
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.13.0",
@@ -23,7 +24,7 @@
     "@types/dagre-d3": "^0.4.39",
     "@types/graphlib": "^2.1.5",
     "@types/lodash": "^4.14.144",
-    "bra-ket-vue": "^0.2.6",
+    "bra-ket-vue": "^0.4.0",
     "core-js": "^3.4.3",
     "d3-array": "^2.3.2",
     "d3-scale": "^3.2.1",
@@ -35,52 +36,51 @@
     "graphlib": "^2.1.7",
     "lodash": "^4.17.19",
     "quantum-tensors": "^0.4.10",
+    "reflect-metadata": "^0.1.13",
     "register-service-worker": "^1.6.2",
     "tone": "^13.8.25",
-    "vue": "^2.6.10",
+    "vue": "^3.0.0-rc.10",
     "vue-confetti": "^2.0.5",
-    "vue-css-donut-chart": "^1.1.6",
-    "vue-gtag": "^1.6.2",
-    "vue-i18n": "^8.21.0",
-    "vue-router": "^3.0.3",
-    "vuex": "^3.0.1",
-    "vuex-class": "^0.3.2"
+    "vue-i18n": "^9.0.0-alpha.17",
+    "vue-router": "^4.0.0-beta.9",
+    "vuex": "^4.0.0-beta.4"
   },
   "devDependencies": {
     "@types/jest": "^24.0.18",
     "@types/markdown-it": "^0.0.9",
-    "@types/webpack-env": "^1.14.1",
     "@typescript-eslint/eslint-plugin": "^4.0.1",
     "@typescript-eslint/parser": "^4.0.1",
-    "@vue/cli-plugin-babel": "^4.1.1",
-    "@vue/cli-plugin-eslint": "^4.1.1",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
     "@vue/cli-plugin-pwa": "^4.1.1",
-    "@vue/cli-plugin-typescript": "^4.1.1",
-    "@vue/cli-service": "^4.1.1",
+    "@vue/cli-plugin-typescript": "^4.5.3",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0-0",
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-standard": "^5.0.0",
-    "@vue/eslint-config-typescript": "^5.0.1",
-    "babel-eslint": "^10.0.1",
+    "@vue/eslint-config-typescript": "^5.0.2",
+    "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-node": "^10.0.0",
     "eslint-plugin-prettier": "^3.1.1",
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.1",
-    "eslint-plugin-vue": "^6.0.1",
-    "eslint-plugin-vue-scoped-css": "^0.4.0",
+    "eslint-plugin-vue": "^7.0.0-beta.3",
+    "eslint-plugin-vue-scoped-css": "^1.0.0",
     "jest": "^24.9.0",
     "last-commit-log": "^3.0.4",
     "markdown-it": "^10.0.0",
     "node-sass": "^4.9.0",
+    "patch-package": "^6.2.2",
     "prettier": "1.19.1",
     "sass-loader": "^8.0.0",
     "ts-node": "^8.8.1",
-    "typescript": "^3.4.3",
-    "vue-class-component": "^7.1.0",
-    "vue-property-decorator": "^8.2.2",
-    "vue-template-compiler": "^2.6.10",
-    "webpack": "^4.41.2",
-    "webpack-env": "^0.8.0"
+    "typescript": "^4.0.2",
+    "vue-class-component": "^8.0.0-alpha.6",
+    "vue-property-decorator": "^8.4.2"
+  },
+  "resolutions": {
+    "vue-property-decorator/vue-class-component": "^8.0.0-alpha.6"
   }
 }
