@@ -1,5 +1,5 @@
 <template>
-  <div slot="left" class="element-list" :class="{ entriesExpanded: entriesExpanded }">
+  <div class="element-list" :class="{ entriesExpanded: entriesExpanded }">
     <h3 class="upper-border" @click="toggleEntries">{{ title }}</h3>
     <router-link v-for="entry in entryList" :key="entry.name" :to="`/info/${entry.name}`">
       <div :class="{ 'not-ready': !entry.ready }">{{ spacedEntry(entry.name) }}</div>
@@ -8,17 +8,17 @@
 </template>
 
 <script lang="ts">
-import { Vue, Component, Prop } from 'vue-property-decorator'
+import { Vue } from 'vue-class-component'
+import { Prop } from 'vue-property-decorator'
 import { IEntryList } from '@/engine/interfaces'
 
-@Component
 export default class EncyclopediaLinkList extends Vue {
   @Prop({ default: 'elements' }) readonly title!: string
   @Prop() readonly entryList!: IEntryList[]
   entriesExpanded = false
 
   toggleEntries(): void {
-    this.$data.entriesExpanded = !this.$data.entriesExpanded
+    this.entriesExpanded = !this.entriesExpanded
   }
 
   spacedEntry(name: string): string {
