@@ -25,7 +25,7 @@
           :max-steps="2"
           :default-step="2"
           :exact-steps="true"
-          @updateRotation="updateRotation"
+          @update-rotation="updateRotation"
         />
       </div>
     </div>
@@ -33,7 +33,8 @@
 </template>
 
 <script lang="ts">
-import { Vue, Prop, Component } from 'vue-property-decorator'
+import { Vue, Options } from 'vue-class-component'
+import { Prop } from 'vue-property-decorator'
 import { Elem } from '@/engine/interfaces'
 import { Coord, Grid, Cell } from '@/engine/classes'
 import { MatrixViewer } from 'bra-ket-vue'
@@ -46,17 +47,17 @@ interface IXYVec {
   vecDirPol: Vector
 }
 
-@Component({
+@Options({
   components: {
     EncyclopediaBoard,
     MatrixViewer,
   },
 })
 export default class EncyclopediaMatrixBoard extends Vue {
-  @Prop({ default: 'Mirror' }) readonly elementName!: string
+  @Prop({ default: 'Mirror' }) readonly elementName!: Elem
   @Prop({ default: () => 10 }) readonly maxSteps!: number
   @Prop({ default: () => 2 }) readonly defaultStep!: number
-  @Prop({ default: '0' }) defaultRotation!: number
+  @Prop({ default: 0 }) defaultRotation!: number
 
   rotation: number = this.defaultRotation
   grid: Grid = Grid.emptyGrid(3, 3)
@@ -67,10 +68,6 @@ export default class EncyclopediaMatrixBoard extends Vue {
     posX: 0,
     posY: 1,
     vecDirPol: Vector.indicator([Dimension.direction(), Dimension.polarization()], ['>', 'H']),
-  }
-
-  $refs!: {
-    grid: HTMLElement
   }
 
   created(): void {
