@@ -83,28 +83,38 @@
 </template>
 
 <script lang="ts">
-import { Vue, Component, Prop } from 'vue-property-decorator'
-import { State, namespace } from 'vuex-class'
-import { GameStateEnum } from '@/engine/interfaces'
+import { Options, setup, Vue } from 'vue-class-component'
+import { Prop } from 'vue-property-decorator'
 import { IStyle } from '@/types'
-import { ActionMethod } from 'vuex'
+import { useRouter } from 'vue-router'
+import { validateInfoPayload } from '@/mixins/gameInterfaces'
+import { storeNamespace } from '@/store'
 
-const userModule = namespace('userModule')
-const optionsModule = namespace('optionsModule')
+const game = storeNamespace('game')
+const user = storeNamespace('user')
+const options = storeNamespace('options')
 
-@Component
+@Options({
+  emits: {
+    hover: validateInfoPayload,
+    'loaded-level': null,
+    play: null,
+    'download-level': null,
+  },
+})
 export default class GameControls extends Vue {
   // FIXME: Can somehow accelerate photon speed by spamming play
   @Prop() readonly frameIndex!: number
   @Prop() readonly totalFrames!: number
   @Prop({ default: '' }) readonly displayStatus!: string
-  @State('gameState') gameState!: GameStateEnum
-  @State('simulationState') simulationState!: boolean
-  @userModule.Action('SAVE_LEVEL') actionSaveLevel!: ActionMethod
-  @userModule.Action('UPDATE_LEVEL') actionUpdateLevel!: ActionMethod
-  @optionsModule.Action('TOGGLE_SOUND') toggleSound!: ActionMethod
-  @userModule.Getter('isLoggedIn') moduleGetterIsLoggedIn!: boolean
-  @optionsModule.Getter('soundActive') soundActive!: boolean
+  gameState = setup(() => game.useState('gameState'))
+  simulationState = setup(() => game.useState('simulationState'))
+  actionSaveLevel = setup(() => user.useAction('SAVE_LEVEL'))
+  actionUpdateLevel = setup(() => user.useAction('UPDATE_LEVEL'))
+  toggleSound = setup(() => options.useAction('TOGGLE_SOUND'))
+  moduleGetterIsLoggedIn = setup(() => user.useGetter('isLoggedIn'))
+  soundActive = setup(() => options.useGetter('soundActive'))
+  router = setup(useRouter)
 
   loadJsonLevelFromFile(event: Event): void {
     const reader = new FileReader()
@@ -261,18 +271,18 @@ export default class GameControls extends Vue {
 
   handleAccount(): void {
     if (!this.moduleGetterIsLoggedIn) {
-      this.$router.push('/login')
+      this.router.push('/login')
     } else {
-      this.$router.push('/myaccount')
+      this.router.push('/myaccount')
     }
   }
 
   handleOptions(): void {
-    this.$router.push('/options')
+    this.router.push('/options')
   }
 
   handleMap(): void {
-    this.$router.push('/levels')
+    this.router.push('/levels')
   }
 
   showSoundHint(): void {
