@@ -12,24 +12,28 @@
 </template>
 
 <script lang="ts">
-import { Vue, Component } from 'vue-property-decorator'
-import { State } from 'vuex-class'
+import { Vue, Options, setup } from 'vue-class-component'
 import { IDialogue } from '@/mixins/dataInterfaces'
 import { getRockTalkById } from './loadRockTalks'
 import RockTalkLine from '@/components/RockTalkPage/RockTalkLine.vue'
 import AppButton from '@/components/AppButton.vue'
+import { useRoute } from 'vue-router'
+import { storeNamespace } from '@/store'
 
-@Component({
+const game = storeNamespace('game')
+
+@Options({
   components: {
     RockTalkLine,
     AppButton,
   },
 })
 export default class InterLevelOverlay extends Vue {
-  @State('currentLevelID') currentLevelID!: string
+  currentLevelID = setup(() => game.useState('currentLevelID'))
+  route = setup(useRoute)
 
   get rockTalkId(): string {
-    return this.$route.params.id
+    return this.route.params.id as string
   }
 
   get rockTalk(): IDialogue {
@@ -41,7 +45,7 @@ export default class InterLevelOverlay extends Vue {
    * @returns a router link :to string parameter
    */
   get nextLevel(): string {
-    return `/level/${parseInt(this.currentLevelID) + 1}`
+    return `/level/${this.currentLevelID + 1}`
   }
 }
 </script>
