@@ -1,4 +1,9 @@
-declare module 'vue-confetti'
-declare module 'vue-css-donut-chart'
+declare module 'vue-confetti' {
+  class Confetti {
+    start(params: Record<string, unknown>): void
+    stop(): void
+  }
+  export { Confetti }
+}
 declare module 'tone'
 // declare module 'quantum-tensors'
